factor out token methods into common package
