+++ pkg/grpc/authn/servicecerttoken/extractor.go
@@ -8,7 +8,7 @@ import (
 	"github.com/pkg/errors"
 	"github.com/stackrox/rox/pkg/grpc/authn"
 	"github.com/stackrox/rox/pkg/grpc/authn/service"
-	"github.com/stackrox/rox/pkg/grpc/client/authn/servicecerttoken"
+	"github.com/stackrox/rox/pkg/grpc/common/authn/servicecerttoken"
 	"github.com/stackrox/rox/pkg/grpc/requestinfo"
 	"github.com/stackrox/rox/pkg/logging"
 	"github.com/stackrox/rox/pkg/mtls"
@@ -30,7 +30,7 @@ func (e extractor) IdentityForRequest(ctx context.Context, ri requestinfo.Reques
 		return nil, nil
 	}
 
-	cert, err := parseToken(token, e.maxLeeway)
+	cert, err := servicecerttoken.ParseToken(token, e.maxLeeway)
 	if err != nil {
 		log.Warnf("Could not parse service cert token: %v", err)
 		return nil, errors.New("could not parse service cert token")
