+++ pkg/grpc/common/authn/servicecerttoken/common.go
@@ -7,13 +7,13 @@ import (
 )
 
 const (
-	// HashAlgo is the cryptographic hash algorithm to use for signatures.
-	HashAlgo = tls.SHA256
+	// hashAlgo is the cryptographic hash algorithm to use for signatures.
+	hashAlgo = tls.SHA256
 	// TokenType is the prefix in the authorization header to identify the token type.
 	TokenType = "ServiceCert"
 )
 
 var (
-	// TokenB64Enc is the base64 encoding flavor to use for tokens.
-	TokenB64Enc = base64.RawStdEncoding
+	// b64Enc is the base64 encoding flavor to use for tokens.
+	b64Enc = base64.RawStdEncoding
 )
